pyproject.toml
[tool.uv.workspace]
members = ["bifrost", "broker", "shared", "dev/kernel-learning"]

[project]
name = "gpu-research-tools"
version = "0.1.0"
description = "Research workspace for GPU provisioning and remote execution tools"
requires-python = ">=3.10"

# Core workspace members
dependencies = [
    "broker",
    "bifrost",
    "shared"
]

[project.optional-dependencies]
# Development experiments (unpublished)
dev-jax-basic = [
    "jax[cuda12] ; sys_platform == 'linux'",
    "jax ; sys_platform == 'darwin'",
]

# Corpus proximity experiment
dev-corpus-proximity = [
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "sentence-transformers>=3.0.0",
    "datasets>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "dacite>=1.8.0",
    "openai>=1.0.0",
    "transformers>=4.30.0",
    "umap-learn>=0.5.5",
    "hdbscan>=0.8.33",
    "syntok>=1.4.4",
]

# Optimizer visualizations
dev-optimizer-visuals = [
    "jax[cuda12] ; sys_platform == 'linux'",
    "jax ; sys_platform == 'darwin'",
    "matplotlib>=3.0.0",
    "streamlit>=1.28.0",
    "plotext>=5.0.0",
]

# Outlier features analysis for MoE models
# Includes bleeding-edge transformers for GPT-OSS & GLM-4 MoE support
dev-outlier-features = [
    "broker",
    "bifrost",
    "shared",
    "nnsight>=0.4",
    "transformers @ git+https://github.com/huggingface/transformers.git@522b79a346c4a188fd16548a7e0bc0ad15b86e5a",  # GPT-OSS & GLM-4 MoE support (pinned commit)
    "huggingface-hub==1.0.0rc2",  # Required by bleeding-edge transformers (pre-release, pinned exact)
    "torch>=2.4.0",
    "accelerate>=0.20.0",
    "datasets>=4.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "triton>=3.4.0; sys_platform == 'linux'",  # MXFP4 quantization support for GPT-OSS
    "kernels; sys_platform == 'linux'",  # Optimized MXFP4 kernels from PyPI for GPT-OSS
]

# LeetGPU CLI for GPU kernel learning challenges
dev-kernel-learning = [
    "kernel-learning",
]

# Nano-inference JAX GPT-2 implementation
dev-nano-inference-jax = [
    "jax[cuda12] ; sys_platform == 'linux'",
    "jax ; sys_platform == 'darwin'",
    "numpy>=1.24.0",
    "transformers>=4.30.0",
    "huggingface-hub>=0.20.0",
    "safetensors>=0.4.0",
    "einops>=0.6.0",
    "torch>=2.0.0",  # Required for comparison utilities
]

# GPT-2 speedrun training
dev-speedrun = [
    "torch>=2.4.0",
    "huggingface-hub>=0.20.0",
    "numpy>=1.24.0",
    "kernels; sys_platform == 'linux'",
]

[project.scripts]
corpus-proximity = "corpus_proximity_cli:main"

[tool.uv.sources]
broker = { workspace = true }
bifrost = { workspace = true }
shared = { workspace = true }
kernel-learning = { workspace = true }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ty>=0.0.1a22",
]
dev-outlier-features = [
    "matplotlib>=3.10.7",
    "plotext>=5.3.2",
    "rich>=14.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shared*"]

[tool.setuptools]
py-modules = ["corpus_proximity_cli"]

[tool.ty.src]
exclude = [
    "**/originals/**",  # Exclude all third-party code in originals/ directories
    "dev/speedrun/history/**",  # Exclude historical speedrun versions
    "dev/speedrun/single-file.py",  # Speedrun example with type issues
    "dev/speedrun/single-file-smoke.py",  # Speedrun example with type issues
    "dev/speedrun/train/00_test_deployment.py",  # Training deployment test
    "dev/nano-inference/basic_blocks.py",  # Incomplete/skeleton code
]
